Add left-clicky and rewrite transparency support to use it
