+++ main.js
@@ -17,51 +17,6 @@ if (isDev) {
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let mainWindow
-let ignoringMouseEvents = false
-let interval
-
-function enableTransparencyChecking() {
-  clearInterval(interval)
-  interval = setInterval(() => {
-    const cursorPoint = electron.screen.getCursorScreenPoint()
-    const windowBounds = mainWindow.getBounds()
-
-    const cursorWithinBounds =
-      (cursorPoint.x >= windowBounds.x && cursorPoint.x <= (windowBounds.x + windowBounds.width)) &&
-      (cursorPoint.y >= windowBounds.y && cursorPoint.y <= (windowBounds.y + windowBounds.height))
-
-    if (cursorWithinBounds) {
-      mainWindow.webContents.capturePage({
-        x: cursorPoint.x - windowBounds.x,
-        y: cursorPoint.y - windowBounds.y,
-        width: 1,
-        height: 1
-      }, (image) => {
-        if (!mainWindow) {
-          return;
-        }
-
-        const buffer = image.getBitmap()
-
-        if (buffer[3] && buffer[3] > 0) {
-          if (ignoringMouseEvents) {
-            mainWindow.setIgnoreMouseEvents(false)
-            ignoringMouseEvents = false
-          }
-        } else {
-          if (!ignoringMouseEvents) {
-            mainWindow.setIgnoreMouseEvents(true)
-            ignoringMouseEvents = true
-          }
-        }
-      })
-    }
-  }, 100)
-}
-
-function disableTransparencyChecking() {
-  clearInterval(interval)
-}
 
 function createWindow () {
   /**
@@ -127,25 +82,13 @@ function createWindow () {
     slashes: true
   }))
 
-  mainWindow.on('restore', () => {
-    enableTransparencyChecking()
-  })
-
-  mainWindow.on('minimize', () => {
-    disableTransparencyChecking()
-  })
-
   // and show window once it's ready (to prevent flashing)
   mainWindow.once('ready-to-show', () => {
     mainWindow.show()
-
-    enableTransparencyChecking()
   })
 
   // Emitted when the window is closed.
   mainWindow.on('closed', function () {
-    disableTransparencyChecking()
-
     // Dereference the window object, usually you would store windows
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
